refactor(forms): Break logic.ts into separate files
This would make it easier to navigate
