+++ packages/forms/signals/src/api/logic/validate.ts
@@ -1,0 +1,38 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.dev/license
+ */
+
+import {addDefaultField} from '../../field/validation';
+import {FieldPathNode} from '../../schema/path_node';
+import {assertPathIsCurrent} from '../../schema/schema';
+import type {FieldContext, SchemaPath, FieldValidator, PathKind, SchemaPathRules} from '../types';
+import {ensureCustomValidationResult} from '../validators/util';
+
+/**
+ * Adds logic to a field to determine if the field has validation errors.
+ *
+ * @param path The target path to add the validation logic to.
+ * @param logic A `Validator` that returns the current validation errors.
+ * @template TValue The type of value stored in the field the logic is bound to.
+ * @template TPathKind The kind of path the logic is bound to (a root path, child path, or item of an array)
+ *
+ * @category logic
+ * @experimental 21.0.0
+ */
+export function validate<TValue, TPathKind extends PathKind = PathKind.Root>(
+  path: SchemaPath<TValue, SchemaPathRules.Supported, TPathKind>,
+  logic: NoInfer<FieldValidator<TValue, TPathKind>>,
+): void {
+  assertPathIsCurrent(path);
+
+  const pathNode = FieldPathNode.unwrapFieldPath(path);
+  pathNode.builder.addSyncErrorRule((ctx) => {
+    return ensureCustomValidationResult(
+      addDefaultField(logic(ctx as FieldContext<TValue, TPathKind>), ctx.field),
+    );
+  });
+}
