+++ packages/forms/signals/src/api/validators/min_length.ts
@@ -7,8 +7,8 @@
  */
 
 import {computed} from '@angular/core';
-import {aggregateMetadata, metadata, validate} from '../logic';
-import {MIN_LENGTH} from '../metadata';
+import {aggregateMetadata, validate} from '../logic';
+import {metadata, MIN_LENGTH} from '../metadata';
 import {SchemaPath, LogicFn, PathKind, SchemaPathRules} from '../types';
 import {minLengthError} from '../validation_errors';
 import {
