+++ packages/forms/signals/src/api/validators/required.ts
@@ -7,8 +7,8 @@
  */
 
 import {computed} from '@angular/core';
-import {aggregateMetadata, metadata, validate} from '../logic';
-import {REQUIRED} from '../metadata';
+import {aggregateMetadata, validate} from '../logic';
+import {metadata, REQUIRED} from '../metadata';
 import {SchemaPath, LogicFn, PathKind, SchemaPathRules} from '../types';
 import {requiredError} from '../validation_errors';
 import {BaseValidatorConfig, getOption, isEmpty} from './util';
